# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
    "/rentals/states": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "description": "Returns a list of Australian states and territories, ordered alphabetically.",
        "responses": {
          "200": {
            "description": "An array of state initialisms.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "examples": [
                    [
                      "ACT",
                      "NSW",
                      "NT",
                      "Qld",
                      "SA",
                      "Tas",
                      "Vic",
                      "WA"
                    ]
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid query parameters: yourParamHere. Query parameters are not permitted."
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/rentals/property-types": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "description": "Returns a list of rental property types, ordered alphabetically.",
        "responses": {
          "200": {
            "description": "An array of property types.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "examples": [
                    [
                      "acreage/semi-rural",
                      "apartment",
                      "duplex/semi-detached",
                      "flat",
                      "house",
                      "other",
                      "studio",
                      "terrace",
                      "townhouse",
                      "unit",
                      "villa"
                    ]
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid query parameters: yourParamHere. Query parameters are not permitted."
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/rentals/{id}": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "description": "Returns an object containing data for the specified rental property. The path parameter {id} is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the rental property",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object containing title, rent, description, propertyType, locality, latitude, longitude, postcode, state, streetAddress, suburb, bathrooms, bedrooms, parkingSpaces, agencyName, amenities, averageRating and numRatings propertiers for the queried rental.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "examples": [
                        "Great location, large family home"
                      ]
                    },
                    "rent": {
                      "type": "integer",
                      "examples": [
                        870
                      ]
                    },
                    "description": {
                      "type": "string",
                      "examples": [
                        "Tucked away in a quiet cul-de-sac in sought-after Eatons Hill, this light-filled home offers relaxed, easy living in a peaceful setting. Elevated to capture breezes year-round, it feels private, airy and welcoming from the moment you arrive. Designed for effortless indoor-outdoor flow, it's perfect for entertaining or unwinding by the pool. Enjoy the convenience of being within walking distance to buses, Eatons Hill Hotel and local shops. A spacious family home in a friendly, well-connected neighbourhood.<br/><br/>Features:<br/>• King-sized master bedroom with spa ensuite and walk-in robe<br/>• Three queen-sized bedrooms with built-in robes<br/>• Three separate internal living areas<br/>• Large dining area plus breakfast bar<br/>• Air conditioning for year-round comfort<br/>• Insulated undercover outdoor entertaining area<br/>• Solar-heated saltwater pool with feature lighting<br/>• Fully fenced, low-maintenance yard<br/>• Generous storage throughout<br/>• Elevated 795sqm block capturing natural light and breezes<br/>• Walking distance to public transport, Eatons Hill Hotel and speciality shops<br/>• Located within highly regarded school catchments and private school bus routes<br/><br/>This property will not last long. Please note that only those that have viewed the property will have their applications processed.<br/><br/>Information contained on any marketing material, website or other portal should not be relied upon and you should make your own enquiries and seek your own independent advice with respect to any property advertised or the information about the property."
                      ]
                    },
                    "propertyType": {
                      "type": "string",
                      "examples": [
                        "house"
                      ]
                    },
                    "locality": {
                      "type": "string",
                      "examples": [
                        "Eatons Hill"
                      ]
                    },
                    "latitude": {
                      "type": "number",
                      "examples": [
                        -27.33943255
                      ]
                    },
                    "longitude": {
                      "type": "number",
                      "examples": [
                        152.94730591
                      ]
                    },
                    "postcode": {
                      "type": "integer",
                      "examples": [
                        4037
                      ]
                    },
                    "state": {
                      "type": "string",
                      "examples": [
                        "Qld"
                      ]
                    },
                    "streetAddress": {
                      "type": "string",
                      "examples": [
                        "23 Valencia Court"
                      ]
                    },
                    "suburb": {
                      "type": "string",
                      "examples": [
                        "Eatons Hill"
                      ]
                    },
                    "bathrooms": {
                      "type": "integer",
                      "examples": [
                        2
                      ]
                    },
                    "bedrooms": {
                      "type": "integer",
                      "examples": [
                        4
                      ]
                    },
                    "parkingSpaces": {
                      "type": "integer",
                      "examples": [
                        2
                      ]
                    },
                    "agencyName": {
                      "type": "string",
                      "examples": [
                        "McGrath Estate Agents Aspley"
                      ]
                    },
                    "amenities": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "Garage: 2, Air Conditioning"
                      ]
                    },
                    "averageRating": {
                      "type": [
                        "number",
                        "null"
                      ],
                      "examples": [
                        4.25
                      ]
                    },
                    "numRatings": {
                      "type": "integer",
                      "examples": [
                        4
                      ]
                    },
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "rating",
                          "user",
                          "dateTime"
                        ],
                        "properties": {
                          "rating": {
                            "type": "integer"
                          },
                          "user": {
                            "type": "string"
                          },
                          "comment": {
                            "type": "string"
                          },
                          "dateTime": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "rating": 4,
                          "user": "john.doe@example.com",
                          "comment": "Comfortable stay with good facilities. A few minor issues, but I would consider staying again.",
                          "dateTime": "2026-03-13T21:51:48.752Z"
                        },
                        {
                          "rating": 5,
                          "user": "jane.doe@example.com",
                          "comment": "Excellent, especially for the price.",
                          "dateTime": "2026-03-29T17:36:15.382Z"
                        },
                        {
                          "rating": 4,
                          "user": "t.chappell@example.com",
                          "dateTime": "2026-04-12T08:19:28.374Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid query parameters: yourParamHere. Query parameters are not permitted."
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The requested rental property could not be found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "No rental exists with this ID."
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}